backend/requirements.txt
fastapi
uvicorn
python-multipart
numpy==2.4.0
pypdf==3.17.4
reportlab==4.0.9
pdf2docx==0.5.6
pdfminer.six==20221105
mammoth==1.6.0
xhtml2pdf==0.2.12